backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
sse-starlette==2.2.1
huggingface-hub==0.28.1
llama-cpp-python==0.3.8
